Fix StackOverflow when printing Vars with meta
Since Vars now may have their reach and use sets (of Vars) as metadata
there is a trivial StackOverflow when printing the metadata with
metadata. The particular fix for this fault is to simply suppress the
two "magic" implementation internal kvs. It's a bit hacky but I don't
see another good way to do this.